gemini/hello_setup.py
from google import genai
import os
from dotenv import load_dotenv

# Load environment variables from .env file
# By default, it looks for a .env file in the current directory
# You can also specify a path: load_dotenv(dotenv_path='path/to/.env')
load_dotenv()
api_key = os.getenv('GEMINI_API_KEY')
model = os.getenv('GEMINI_MODEL')
# The client gets the API key from the environment variable `GEMINI_API_KEY`.
client = genai.Client()

response = client.models.generate_content(
    model=model, 
    contents="Explain how AI works in a few words"
)
print(response.text)